refact(api): split schemas_common.py into 7 per-domain schema modules (#5799)
Moves 370 of 375 classes out of the 3,748-line monolith into domain files:
  schemas_terminal.py   (41 classes) - Terminal*, AgentTerminal*, SSH*
  schemas_analytics.py  (32 classes) - Analytics*, Cost*, Budget*, Usage*, Metrics*
  schemas_knowledge.py  (82 classes) - Knowledge*
  schemas_agent.py      (55 classes) - Agent*, Memory*, LLM*
  schemas_system.py     (34 classes) - System*, NPU*, WakeWord*, FeatureFlag*
  schemas_workflows.py  (57 classes) - Workflow*, Registry*, RUM*, Elevation*, AdvancedControl*
  schemas_code.py       (69 classes) - CodeReview*, Git*, Skills*, Database*, Template*, etc.
  schemas_common.py     (5 classes)  - DataResponse, SuccessResponse, SuccessMessageResponse,
                                       SuccessDataResponse, UsageRecordResponse (cross-domain)

Updates 48 API files to import from their domain-specific schema module.
All 56 modified/new files pass py_compile.
