+++ autobot-backend/api/knowledge_ai_stack.py
@@ -31,7 +31,6 @@ from utils.response_helpers import (
     handle_ai_stack_error,
 )
 
-from .schemas_common import DataResponse
 from api.schemas_common import DataResponse, SuccessResponse
 
 logger = logging.getLogger(__name__)
