+++ autobot-backend/api/knowledge_ownership.py
@@ -15,7 +15,7 @@ from typing import Optional
 from fastapi import APIRouter, Depends, HTTPException, Request
 
 from api.knowledge_models import ShareFactRequest, UpdateVisibilityRequest
-from api.schemas_common import (
+from api.schemas_knowledge import (
     KnowledgeMyFactsResponse,
     KnowledgeShareFactResponse,
     KnowledgeSharedWithMeResponse,
