+++ autobot-backend/api/llm.py
@@ -7,7 +7,7 @@ import logging
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import JSONResponse
 
-from api.schemas_common import (
+from api.schemas_agent import (
     LLMConnectionTestResponse,
     LLMCurrentResponse,
     LLMEmbeddingModelsResponse,
