+++ autobot-backend/api/schemas_common.py
@@ -2,14 +2,17 @@
 # Copyright (c) 2025 mrveiss
 # Author: mrveiss
 """
-Shared Pydantic response schemas for AutoBot API endpoints.
+Shared cross-domain Pydantic response schemas for AutoBot API endpoints.
 
-Provides typed response models for terminal.py and terminal_tools.py.
-All models are used as response_model= on FastAPI router decorators so that
-FastAPI generates accurate OpenAPI docs and validates outbound data.
-
-Schema naming convention:
-  <Domain><EntityOrAction>Response
+These are truly generic types used across multiple unrelated domains.
+Domain-specific schemas live in:
+  schemas_terminal.py    - Terminal*, AgentTerminal*, SSH*, etc.
+  schemas_analytics.py   - Analytics*, Cost*, Budget*, Usage*, Metrics*, etc.
+  schemas_knowledge.py   - Knowledge* schemas
+  schemas_agent.py       - Agent*, Memory*, LLM* schemas
+  schemas_system.py      - System*, NPU*, WakeWord*, FeatureFlag*, etc.
+  schemas_workflows.py   - Workflow*, Registry*, RUM*, Elevation*, AdvancedControl*, etc.
+  schemas_code.py        - CodeReview*, Git*, Skills*, Database*, Template*, etc.
 """
 
 from typing import Any, Dict, List, Optional
@@ -18,9 +21,8 @@ from pydantic import BaseModel
 
 
 # ---------------------------------------------------------------------------
-# Generic / reusable
+# Generic / reusable (used across multiple unrelated domains)
 # ---------------------------------------------------------------------------
-
 
 class SuccessMessageResponse(BaseModel):
     """Base for endpoints that always return success: bool + message: str."""
@@ -29,12 +31,14 @@ class SuccessMessageResponse(BaseModel):
     message: str
 
 
+
 class SuccessDataResponse(BaseModel):
     """Base for endpoints that return success, message, and an optional data dict."""
 
     success: bool
     message: str
     data: Optional[Dict[str, Any]] = None
+
 
 
 class SuccessResponse(BaseModel):
@@ -48,299 +52,6 @@ class SuccessResponse(BaseModel):
 # Terminal session schemas  (terminal.py — session lifecycle)
 # ---------------------------------------------------------------------------
 
-
-class TerminalSessionCreateResponse(BaseModel):
-    """Response for POST /sessions."""
-
-    session_id: str
-    status: str
-    security_level: str
-    websocket_url: str
-    created_at: str
-    ssh_keys: Optional[Dict[str, Any]] = None
-
-
-class TerminalSessionItem(BaseModel):
-    """Single session entry within a list response."""
-
-    session_id: str
-    user_id: Optional[str] = None
-    security_level: Optional[str] = None
-    created_at: Optional[str] = None
-    is_active: bool
-
-
-class TerminalSessionListResponse(BaseModel):
-    """Response for GET /sessions."""
-
-    sessions: List[TerminalSessionItem]
-    total: int
-    active: int
-
-
-class TerminalSessionDetailResponse(BaseModel):
-    """Response for GET /sessions/{session_id}."""
-
-    session_id: str
-    config: Dict[str, Any]
-    is_active: bool
-    statistics: Dict[str, Any]
-
-
-class TerminalSessionDeleteResponse(BaseModel):
-    """Response for DELETE /sessions/{session_id}."""
-
-    session_id: str
-    status: str
-
-
-# ---------------------------------------------------------------------------
-# SSH key management schemas  (terminal.py — Issue #211)
-# ---------------------------------------------------------------------------
-
-
-class SSHKeyItem(BaseModel):
-    """A single SSH key entry as returned by get_session_keys()."""
-
-    name: str
-    fingerprint: Optional[str] = None
-    has_passphrase: Optional[bool] = None
-    key_path: Optional[str] = None
-
-
-class SSHKeyListResponse(BaseModel):
-    """Response for GET /sessions/{session_id}/ssh-keys."""
-
-    session_id: str
-    keys: List[Any]  # key dicts are opaque — shape defined by terminal_secrets_service
-    total: int
-
-
-class SSHKeyAgentResponse(BaseModel):
-    """Response for POST /sessions/{session_id}/ssh-keys/{key_name}/agent."""
-
-    session_id: str
-    key_name: str
-    status: str
-    message: str
-
-
-class SSHKeyPathResponse(BaseModel):
-    """Response for GET /sessions/{session_id}/ssh-keys/{key_name}/path."""
-
-    session_id: str
-    key_name: str
-    key_path: str
-    usage: str
-
-
-# ---------------------------------------------------------------------------
-# Command / input / signal schemas
-# ---------------------------------------------------------------------------
-
-
-class CommandAssessResponse(BaseModel):
-    """Response for POST /command."""
-
-    command: str
-    risk_level: str
-    status: str
-    message: str
-    requires_confirmation: bool
-
-
-class TerminalInputResponse(BaseModel):
-    """Response for POST /sessions/{session_id}/input."""
-
-    session_id: str
-    status: str
-    input: str
-
-
-class TerminalSignalResponse(BaseModel):
-    """Response for POST /sessions/{session_id}/signal/{signal_name}."""
-
-    session_id: str
-    signal: str
-    status: str
-
-
-# ---------------------------------------------------------------------------
-# History / audit schemas
-# ---------------------------------------------------------------------------
-
-
-class TerminalCommandHistoryResponse(BaseModel):
-    """Response for GET /sessions/{session_id}/history."""
-
-    session_id: str
-    is_active: bool
-    history: List[Any]
-    total_commands: Optional[int] = None
-    message: Optional[str] = None
-
-
-class TerminalAuditLogResponse(BaseModel):
-    """Response for GET /audit/{session_id}."""
-
-    session_id: str
-    audit_available: bool
-    security_level: Optional[str] = None
-    message: str
-
-
-# ---------------------------------------------------------------------------
-# Info / health / status schemas
-# ---------------------------------------------------------------------------
-
-
-class TerminalInfoResponse(BaseModel):
-    """Response for GET /."""
-
-    name: str
-    version: str
-    description: str
-    features: List[str]
-    endpoints: Dict[str, str]
-    security_levels: List[str]
-    notice: str
-
-
-class TerminalHealthComponents(BaseModel):
-    terminal_manager: str
-    websocket_manager: str
-    pty_system: str
-    session_manager: str
-
-
-class TerminalHealthMetrics(BaseModel):
-    active_sessions: int
-    manager_initialized: bool
-
-
-class TerminalHealthResponse(BaseModel):
-    """Response for GET /health (healthy path)."""
-
-    status: str
-    service: str
-    components: Optional[TerminalHealthComponents] = None
-    metrics: Optional[TerminalHealthMetrics] = None
-    error: Optional[str] = None
-
-
-class TerminalStatusFeatures(BaseModel):
-    pty_support: bool
-    websocket_support: bool
-    command_validation: bool
-    security_policies: bool
-    approval_workflow: bool
-    agent_integration: bool
-
-
-class TerminalStatusSessionInfo(BaseModel):
-    active_sessions: int
-    max_concurrent_sessions: Optional[int] = None
-
-
-class TerminalSystemStatusResponse(BaseModel):
-    """Response for GET /status."""
-
-    status: str
-    terminal_types: Optional[List[str]] = None
-    features: Optional[TerminalStatusFeatures] = None
-    session_info: Optional[TerminalStatusSessionInfo] = None
-    error: Optional[str] = None
-
-
-class TerminalCapabilitiesResponse(BaseModel):
-    """Response for GET /capabilities."""
-
-    pty_management: bool
-    websocket_streaming: bool
-    command_execution: bool
-    security_validation: bool
-    session_management: bool
-    terminal_types: Dict[str, Any]
-    pty_features: Dict[str, Any]
-    websocket_features: Dict[str, Any]
-
-
-class TerminalSecurityPoliciesResponse(BaseModel):
-    """Response for GET /security."""
-
-    command_validation: str
-    risk_assessment: str
-    risk_levels: Dict[str, str]
-    security_executor: str
-    approval_workflow: Dict[str, Any]
-    audit_logging: Dict[str, Any]
-
-
-class TerminalImplementationItem(BaseModel):
-    name: str
-    description: str
-    frontend_component: str
-    backend_api: str
-    approval_workflow: bool
-    service_layer: Optional[str] = None
-
-
-class TerminalFeaturesResponse(BaseModel):
-    """Response for GET /features."""
-
-    manager_class: str
-    websocket_class: str
-    pty_implementation: str
-    implementations: List[TerminalImplementationItem]
-    features: Dict[str, str]
-    shared_infrastructure: Dict[str, str]
-
-
-# ---------------------------------------------------------------------------
-# Stats schema
-# ---------------------------------------------------------------------------
-
-
-class TerminalStatsResponse(BaseModel):
-    """Response for GET /stats.
-
-    The session manager returns an opaque dict whose shape varies between
-    the 'all sessions' and 'single session' cases.  A permissive model
-    lets FastAPI pass the dict through while still documenting the endpoint.
-    """
-
-    model_config = {"extra": "allow"}
-
-
-# ---------------------------------------------------------------------------
-# Admin execute schema  (SLM admin terminal — Issue #983)
-# ---------------------------------------------------------------------------
-
-
-class AdminExecuteResponse(BaseModel):
-    """Response for POST /execute."""
-
-    stdout: str
-    stderr: str
-    exit_code: int
-
-
-# ---------------------------------------------------------------------------
-# terminal_tools.py schemas
-# ---------------------------------------------------------------------------
-
-
-class PackageManagersResponse(BaseModel):
-    """Response for GET /terminal/package-managers."""
-
-    detected: Optional[str] = None
-    available: List[str]
-    package_managers: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# Generic data envelope  (ai_stack_integration, agent, etc.)
-# ---------------------------------------------------------------------------
 
 
 class DataResponse(BaseModel):
@@ -361,2900 +72,6 @@ class DataResponse(BaseModel):
 # ---------------------------------------------------------------------------
 
 
-class AgentMessageResponse(BaseModel):
-    """Response for /goal, /pause, /resume — plain {"message": str}."""
-
-    message: str
-
-
-class AgentCommandApprovalResponse(BaseModel):
-    """Response for POST /command_approval."""
-
-    message: str
-    task_id: str
-    approved: bool
-
-
-class AgentCommandExecuteResponse(BaseModel):
-    """Response for POST /execute_command — success path returns {"message", "output", "status"}."""
-
-    message: str
-    output: Optional[str] = None
-    status: Optional[str] = None
-
-
-class AgentHealthResponse(BaseModel):
-    """Response for GET /health/enhanced."""
-
-    status: str
-    ai_stack_available: bool
-    multi_agent_coordination: bool
-    enhanced_capabilities: bool
-    timestamp: str
-    error: Optional[str] = None
-
-
-# ---------------------------------------------------------------------------
-# enhanced_memory.py schemas
-# ---------------------------------------------------------------------------
-
-
-class MemoryStatisticsResponse(BaseModel):
-    """Response for GET /statistics."""
-
-    period_days: int
-    timestamp: str
-    task_execution: Optional[Any] = None
-    markdown_system: Optional[Any] = None
-    active_tasks: Optional[Any] = None
-    performance_insights: Optional[Any] = None
-
-
-class MemoryTaskHistoryResponse(BaseModel):
-    """Response for GET /tasks/history."""
-
-    total_records: int
-    filter_criteria: Dict[str, Any]
-    tasks: List[Any]
-
-
-class MemoryTaskCreateResponse(BaseModel):
-    """Response for POST /tasks."""
-
-    task_id: str
-    status: str
-    timestamp: str
-
-
-class MemoryTaskUpdateResponse(BaseModel):
-    """Response for PUT /tasks/{task_id}."""
-
-    task_id: str
-    status: str
-    timestamp: str
-
-
-class MemoryMarkdownReferenceResponse(BaseModel):
-    """Response for POST /tasks/{task_id}/markdown-reference."""
-
-    task_id: str
-    markdown_file: str
-    reference_type: str
-    status: str
-    timestamp: str
-
-
-class MemoryMarkdownScanResponse(BaseModel):
-    """Response for GET /markdown/scan."""
-
-    status: str
-    scan_results: Optional[Any] = None
-    timestamp: str
-
-
-class MemoryMarkdownSearchResponse(BaseModel):
-    """Response for GET /markdown/search."""
-
-    query: str
-    filters: Dict[str, Any]
-    total_results: int
-    results: List[Any]
-
-
-class MemoryDocumentReferencesResponse(BaseModel):
-    """Response for GET /markdown/{file_path}/references."""
-
-    file_path: str
-    timestamp: str
-    references: Optional[Any] = None
-
-
-class MemoryEmbeddingCacheStatsResponse(BaseModel):
-    """Response for GET /embeddings/cache-stats."""
-
-    cache_size: Optional[Any] = None
-    timestamp: str
-    status: str
-
-
-class MemoryCleanupResponse(BaseModel):
-    """Response for DELETE /cleanup."""
-
-    status: str
-    cleanup_results: Dict[str, Any]
-    days_kept: int
-    timestamp: str
-
-
-class MemoryActiveTasksResponse(BaseModel):
-    """Response for GET /active-tasks."""
-
-    count: int
-    active_tasks: List[Any]
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# agent_config.py schemas
-# ---------------------------------------------------------------------------
-
-
-class AgentConfigEnableDisableResponse(BaseModel):
-    """Response for POST /agents/{agent_id}/enable and /disable."""
-
-    status: str
-    message: str
-    agent_name: str
-
-
-class AgentConfigUpdateModelResponse(BaseModel):
-    """Response for POST /agents/{agent_id}/model."""
-
-    status: str
-    message: str
-    updated_config: Dict[str, Any]
-
-
-class AgentConfigHealthResponse(BaseModel):
-    """Response for GET /agents/{agent_id}/health."""
-
-    agent_id: str
-    agent_name: str
-    status: str
-    enabled: bool
-    model: Optional[str] = None
-    checks: Dict[str, Any]
-    timestamp: str
-    response_time: float
-
-
-class AgentConfigOverviewResponse(BaseModel):
-    """Response for GET /status/overview."""
-
-    total_agents: int
-    enabled_agents: int
-    healthy_agents: int
-    unhealthy_agents: int
-    disabled_agents: int
-    overall_health: str
-    agents: List[Any]
-    timestamp: str
-
-# ---------------------------------------------------------------------------
-# LLM schemas  (llm.py)
-# ---------------------------------------------------------------------------
-
-
-class LLMConnectionTestResponse(BaseModel):
-    """Response for POST /llm/test_connection."""
-
-    status: str
-    message: Optional[str] = None
-
-    model_config = {"extra": "allow"}
-
-
-class LLMModelsResponse(BaseModel):
-    """Response for GET /llm/models."""
-
-    models: List[Any]
-    total_count: int
-
-
-class LLMCurrentResponse(BaseModel):
-    """Response for GET /llm/current."""
-
-    model: str
-    provider: str
-    config: Dict[str, Any]
-
-
-class LLMEmbeddingModelsResponse(BaseModel):
-    """Response for GET /llm/embedding/models."""
-
-    models: List[Any]
-    total_count: int
-
-
-# ---------------------------------------------------------------------------
-# LLM Optimization schemas  (llm_optimization.py)
-# ---------------------------------------------------------------------------
-
-
-class LLMOptimizationHealthResponse(BaseModel):
-    """Response for GET /llm-optimization/health."""
-
-    status: str
-    available_models: int
-    cache_size: int
-    ollama_connection: bool
-    redis_connected: bool
-
-
-class LLMAvailableModelsResponse(BaseModel):
-    """Response for GET /llm-optimization/models/available."""
-
-    models_count: int
-    models: List[Any]
-    timestamp: float
-
-
-class LLMSelectModelResponse(BaseModel):
-    """Response for POST /llm-optimization/models/select."""
-
-    selected_model: Optional[str] = None
-    model_details: Optional[Dict[str, Any]] = None
-    task_complexity: str
-    selection_reasoning: Dict[str, Any]
-    timestamp: float
-
-
-class LLMTrackPerformanceResponse(BaseModel):
-    """Response for POST /llm-optimization/models/performance/track."""
-
-    status: str
-    message: str
-    recorded_data: Dict[str, Any]
-
-
-class LLMOptimizationSuggestionsResponse(BaseModel):
-    """Response for GET /llm-optimization/optimization/suggestions."""
-
-    suggestions_count: int
-    suggestions: List[Any]
-    timestamp: float
-
-
-class LLMModelsComparisonResponse(BaseModel):
-    """Response for GET /llm-optimization/models/comparison."""
-
-    comparison: Dict[str, Any]
-    best_models: Dict[str, Any]
-    total_models: int
-    timestamp: float
-
-    model_config = {"extra": "allow"}
-
-
-class LLMBenchmarkResponse(BaseModel):
-    """Response for POST /llm-optimization/models/benchmark/{model_name}."""
-
-    model_name: str
-    test_queries: List[str]
-    iterations_per_query: int
-    status: str
-    message: str
-    expected_metrics: List[str]
-    timestamp: float
-
-
-class LLMSystemResourcesResponse(BaseModel):
-    """Response for GET /llm-optimization/system/resources."""
-
-    resources: Dict[str, Any]
-    recommendations: List[Any]
-    optimal_model_size_gb: float
-    timestamp: float
-
-
-class LLMOptimizationConfigResponse(BaseModel):
-    """Response for GET /llm-optimization/config."""
-
-    performance_threshold: float
-    cache_ttl: int
-    min_samples: int
-    model_classification: Dict[str, Any]
-    task_complexity_levels: List[str]
-    optimization_factors: List[str]
-
-
-class LLMInferenceSettingsResponse(BaseModel):
-    """Response for GET/POST /llm-optimization/inference/settings."""
-
-    settings: Dict[str, Any]
-    timestamp: float
-
-
-class LLMInferenceMetricsResponse(BaseModel):
-    """Response for GET /llm-optimization/inference/metrics."""
-
-    optimization: Dict[str, Any]
-    cache: Dict[str, Any]
-    provider_usage: Dict[str, Any]
-    total_requests: int
-    avg_response_time: float
-    fallback_count: int
-    timestamp: float
-
-
-class LLMProviderOptimizationSummaryResponse(BaseModel):
-    """Response for GET /llm-optimization/inference/provider/{provider_type}/optimizations."""
-
-    provider: str
-    is_local: bool
-    is_cloud: bool
-    optimizations: Dict[str, Any]
-    timestamp: float
-
-
-class LLMModelPerformanceHistoryResponse(BaseModel):
-    """Response for GET /llm-optimization/models/performance/history/{model_name}.
-
-    Shape is opaque — defined by model.to_performance_history_dict().
-    """
-
-    model_config = {"extra": "allow"}
-
-
-# ---------------------------------------------------------------------------
-# LLM Awareness schemas  (llm_awareness.py)
-# ---------------------------------------------------------------------------
-
-
-class LLMAwarenessStatusResponse(BaseModel):
-    """Response for GET /llm-awareness/status."""
-
-    status: str
-    service: str
-    timestamp: str
-    system_identity: Dict[str, Any]
-    capabilities_count: int
-    system_maturity: str
-
-
-class LLMSystemContextResponse(BaseModel):
-    """Response for GET /llm-awareness/context."""
-
-    status: str
-    format: str
-    context: Dict[str, Any]
-    timestamp: str
-
-
-class LLMCapabilitiesSummaryResponse(BaseModel):
-    """Response for GET /llm-awareness/capabilities."""
-
-    status: str
-    capabilities: Dict[str, Any]
-    timestamp: str
-
-
-class LLMInjectContextResponse(BaseModel):
-    """Response for POST /llm-awareness/inject-context."""
-
-    status: str
-    original_prompt: str
-    enhanced_prompt: str
-    context_level: str
-    timestamp: str
-
-
-class LLMAnalyzeQueryResponse(BaseModel):
-    """Response for POST /llm-awareness/analyze-query."""
-
-    status: str
-    analysis: Dict[str, Any]
-    timestamp: str
-
-
-class LLMCapabilitySummaryTextResponse(BaseModel):
-    """Response for GET /llm-awareness/summary/text."""
-
-    status: str
-    summary: str
-    format: str
-    timestamp: str
-
-
-class LLMPhaseInfoResponse(BaseModel):
-    """Response for GET /llm-awareness/phase-info."""
-
-    status: str
-    phase_info: Dict[str, Any]
-    timestamp: str
-
-
-class LLMAwarenessMetricsResponse(BaseModel):
-    """Response for GET /llm-awareness/metrics."""
-
-    status: str
-    metrics: Dict[str, Any]
-    timestamp: str
-
-
-class LLMExportAwarenessResponse(BaseModel):
-    """Response for POST /llm-awareness/export."""
-
-    status: str
-    message: str
-    output_path: str
-    format: str
-    timestamp: str
-
-
-class LLMAwarenessHealthResponse(BaseModel):
-    """Response for GET /llm-awareness/health."""
-
-    status: str
-    service: str
-    components: Dict[str, Any]
-    system_maturity: str
-    capabilities_count: int
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# system.py schemas
-# ---------------------------------------------------------------------------
-
-
-class SystemFrontendConfigResponse(BaseModel):
-    """Response for GET /frontend-config."""
-
-    status: str
-    config: Dict[str, Any]
-    timestamp: str
-
-
-class SystemHealthResponse(BaseModel):
-    """Response for GET /health and GET /system/health.
-
-    Shape varies between healthy/degraded/unhealthy paths — extra fields
-    (cpu_percent, memory_percent, services, etc.) are allowed through.
-    """
-
-    model_config = {"extra": "allow"}
-
-    status: str
-    timestamp: str
-
-
-class SystemInfoResponse(BaseModel):
-    """Response for GET /info."""
-
-    name: str
-    version: str
-    python_version: str
-    timestamp: str
-    features: Dict[str, Any]
-
-
-class SystemReloadConfigResponse(BaseModel):
-    """Response for POST /reload_config."""
-
-    status: str
-    message: str
-    timestamp: str
-
-
-class SystemPromptReloadResponse(BaseModel):
-    """Response for GET /prompt_reload."""
-
-    status: str
-    message: str
-    timestamp: str
-
-
-class SystemAdminCheckResponse(BaseModel):
-    """Response for GET /admin_check."""
-
-    user: str
-    admin: bool
-    timestamp: str
-
-
-class SystemDynamicImportResponse(BaseModel):
-    """Response for POST /dynamic_import."""
-
-    status: str
-    message: str
-    module_info: Dict[str, Any]
-    timestamp: str
-
-
-class SystemCacheStatsResponse(BaseModel):
-    """Response for GET /cache/stats.
-
-    Redis info and performance sections are dynamic — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    timestamp: str
-    cache: Dict[str, Any]
-
-
-class SystemCacheActivityResponse(BaseModel):
-    """Response for GET /cache/activity."""
-
-    model_config = {"extra": "allow"}
-
-    timestamp: str
-    activity: Dict[str, Any]
-
-
-class SystemMetricsResponse(BaseModel):
-    """Response for GET /metrics.
-
-    psutil-derived payload is dynamic; extra fields allowed through.
-    """
-
-    model_config = {"extra": "allow"}
-
-    timestamp: str
-
-
-class SystemCacheCoordinatorStatsResponse(BaseModel):
-    """Response for GET /api/cache/stats.
-
-    Shape is defined by CacheCoordinator.get_unified_stats() — opaque.
-    """
-
-    model_config = {"extra": "allow"}
-
-
-class SystemCacheEvictResponse(BaseModel):
-    """Response for POST /api/cache/evict."""
-
-    status: str
-    evicted: int
-    timestamp: str
-
-
-class SystemCacheClearResponse(BaseModel):
-    """Response for POST /api/cache/clear/{cache_name}."""
-
-    status: str
-    cache: str
-    timestamp: str
-
-
-class SystemBackupStatusResponse(BaseModel):
-    """Response for GET /system/backup/status.
-
-    BackupScheduler.get_status() returns an opaque dict with a timestamp added.
-    """
-
-    model_config = {"extra": "allow"}
-
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# agent_terminal.py schemas
-# ---------------------------------------------------------------------------
-
-
-class AgentTerminalSessionCreateResponse(BaseModel):
-    """Response for POST /agent-terminal/sessions."""
-
-    status: str
-    session_id: str
-    agent_id: str
-    agent_role: str
-    conversation_id: Optional[str] = None
-    host: str
-    state: str
-    created_at: float
-    pty_session_id: Optional[str] = None
-
-
-class AgentTerminalSessionItem(BaseModel):
-    """Single session entry within list response."""
-
-    session_id: str
-    agent_id: str
-    agent_role: str
-    conversation_id: Optional[str] = None
-    host: str
-    state: str
-    created_at: float
-    last_activity: Optional[float] = None
-    command_count: int
-    pty_session_id: Optional[str] = None
-
-
-class AgentTerminalSessionListResponse(BaseModel):
-    """Response for GET /agent-terminal/sessions."""
-
-    status: str
-    total: int
-    sessions: List[AgentTerminalSessionItem]
-
-
-class AgentTerminalSessionDetailResponse(BaseModel):
-    """Response for GET /agent-terminal/sessions/{session_id}.
-
-    session_info dict shape is opaque — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    status: str
-
-
-class AgentTerminalSessionDeleteResponse(BaseModel):
-    """Response for DELETE /agent-terminal/sessions/{session_id}."""
-
-    status: str
-    session_id: str
-
-
-class AgentTerminalCommandStateResponse(BaseModel):
-    """Response for GET /agent-terminal/commands/{command_id}."""
-
-    command_id: str
-    terminal_session_id: Optional[str] = None
-    chat_id: Optional[str] = None
-    command: str
-    purpose: Optional[str] = None
-    state: str
-    output: Optional[str] = None
-    stderr: Optional[str] = None
-    return_code: Optional[int] = None
-    risk_level: str
-    risk_reasons: Optional[List[str]] = None
-    requested_at: Optional[float] = None
-    approved_at: Optional[float] = None
-    execution_started_at: Optional[float] = None
-    execution_completed_at: Optional[float] = None
-    approved_by_user_id: Optional[str] = None
-    approval_comment: Optional[str] = None
-
-
-class AgentTerminalInfoResponse(BaseModel):
-    """Response for GET /agent-terminal/."""
-
-    name: str
-    version: str
-    description: str
-    features: List[str]
-    agent_roles: List[str]
-    session_states: List[str]
-    endpoints: Dict[str, str]
-    security_features: Dict[str, str]
-
-
-class AgentTerminalToolApprovalResponse(BaseModel):
-    """Response for POST /agent-terminal/tools/approve/{approval_id}."""
-
-    status: str
-    approval_id: str
-    approved: bool
-
-
-class AgentTerminalHostSelectionRequestResponse(BaseModel):
-    """Response for POST /agent-terminal/host-selection/request."""
-
-    request_id: str
-    status: str
-    message: str
-
-
-class AgentTerminalHostSelectionGetResponse(BaseModel):
-    """Response for GET /agent-terminal/host-selection/{request_id}."""
-
-    request_id: str
-    status: str
-    selected_host_id: Optional[str] = None
-    selected_host_name: Optional[str] = None
-    connection_info: Optional[Dict[str, Any]] = None
-    created_at: str
-    updated_at: Optional[str] = None
-
-
-class AgentTerminalHostSelectionSubmitResponse(BaseModel):
-    """Response for POST /agent-terminal/host-selection/{request_id}/select."""
-
-    status: str
-    request_id: str
-    selected_host_id: Optional[str] = None
-    selected_host_name: Optional[str] = None
-    connection_info: Optional[Dict[str, Any]] = None
-
-
-class AgentTerminalHostSelectionCancelResponse(BaseModel):
-    """Response for POST /agent-terminal/host-selection/{request_id}/cancel."""
-
-    status: str
-    request_id: str
-
-
-class AgentTerminalPendingSelectionsResponse(BaseModel):
-    """Response for GET /agent-terminal/host-selection."""
-
-    status: str
-    pending_count: int
-    pending_selections: List[Dict[str, Any]]
-
-
-# ---------------------------------------------------------------------------
-# knowledge_metadata.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeMetadataTemplateResponse(BaseModel):
-    """Response for POST /metadata/templates and PUT /metadata/templates/{id}."""
-
-    status: str
-    message: Optional[str] = None
-    template: Optional[Dict[str, Any]] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeMetadataTemplateListResponse(BaseModel):
-    """Response for GET /metadata/templates."""
-
-    status: str
-    count: Optional[int] = None
-    templates: Optional[List[Any]] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeMetadataTemplateDetailResponse(BaseModel):
-    """Response for GET /metadata/templates/{template_id}."""
-
-    status: str
-    template: Optional[Dict[str, Any]] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeMetadataTemplateDeleteResponse(BaseModel):
-    """Response for DELETE /metadata/templates/{template_id}."""
-
-    status: str
-    message: Optional[str] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeMetadataValidateResponse(BaseModel):
-    """Response for POST /metadata/validate."""
-
-    valid: Optional[bool] = None
-    errors: Optional[List[Any]] = None
-    warnings: Optional[List[Any]] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeMetadataSearchResponse(BaseModel):
-    """Response for POST /metadata/search."""
-
-    status: str
-    count: Optional[int] = None
-    facts: Optional[List[Any]] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeFactVersionListResponse(BaseModel):
-    """Response for GET /facts/{fact_id}/versions."""
-
-    status: str
-    fact_id: Optional[str] = None
-    versions: Optional[List[Any]] = None
-    count: Optional[int] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeFactVersionDetailResponse(BaseModel):
-    """Response for GET /facts/{fact_id}/versions/{version}."""
-
-    status: str
-    fact_id: Optional[str] = None
-    version: Optional[int] = None
-    content: Optional[Any] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeFactRevertResponse(BaseModel):
-    """Response for POST /facts/{fact_id}/revert."""
-
-    status: str
-    message: Optional[str] = None
-    new_version: Optional[int] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeFactVersionCompareResponse(BaseModel):
-    """Response for POST /facts/{fact_id}/versions/compare."""
-
-    status: str
-    fact_id: Optional[str] = None
-    version_a: Optional[int] = None
-    version_b: Optional[int] = None
-    diff: Optional[Any] = None
-
-    model_config = {"extra": "allow"}
-
-
-class KnowledgeFactVersionHistoryDeleteResponse(BaseModel):
-    """Response for DELETE /facts/{fact_id}/versions."""
-
-    status: str
-    message: Optional[str] = None
-
-    model_config = {"extra": "allow"}
-
-
-# ---------------------------------------------------------------------------
-# knowledge_collections.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeCollectionCreateResponse(BaseModel):
-    """Response for POST /collections."""
-
-    status: str
-    collection: Optional[Dict[str, Any]] = None
-    message: Optional[str] = None
-
-
-class KnowledgeCollectionListResponse(BaseModel):
-    """Response for GET /collections."""
-
-    status: str
-    collections: List[Any]
-    total_count: int
-    returned_count: int
-    limit: int
-    offset: int
-    has_more: bool
-
-
-class KnowledgeCollectionDetailResponse(BaseModel):
-    """Response for GET /collections/{collection_id}."""
-
-    status: str
-    collection: Optional[Dict[str, Any]] = None
-
-
-class KnowledgeCollectionUpdateResponse(BaseModel):
-    """Response for PUT /collections/{collection_id}."""
-
-    status: str
-    collection: Optional[Dict[str, Any]] = None
-    message: Optional[str] = None
-
-
-class KnowledgeCollectionDeleteResponse(BaseModel):
-    """Response for DELETE /collections/{collection_id}."""
-
-    status: str
-    collection_id: Optional[str] = None
-    facts_in_collection: int
-    facts_deleted: int
-    message: Optional[str] = None
-
-
-class KnowledgeCollectionAddFactsResponse(BaseModel):
-    """Response for POST /collections/{collection_id}/facts."""
-
-    status: str
-    collection_id: Optional[str] = None
-    added_count: int
-    already_in_collection: int
-    not_found: List[Any]
-    total_facts: int
-    message: Optional[str] = None
-
-
-class KnowledgeCollectionRemoveFactsResponse(BaseModel):
-    """Response for DELETE /collections/{collection_id}/facts."""
-
-    status: str
-    collection_id: Optional[str] = None
-    removed_count: int
-    not_in_collection: int
-    total_facts: int
-    message: Optional[str] = None
-
-
-class KnowledgeCollectionFactsListResponse(BaseModel):
-    """Response for GET /collections/{collection_id}/facts."""
-
-    status: str
-    collection_id: Optional[str] = None
-    collection_name: Optional[str] = None
-    facts: List[Any]
-    total_count: int
-    returned_count: int
-    limit: int
-    offset: int
-    has_more: bool
-
-
-class KnowledgeFactCollectionsResponse(BaseModel):
-    """Response for GET /facts/{fact_id}/collections."""
-
-    status: str
-    fact_id: Optional[str] = None
-    collections: List[Any]
-    count: int
-
-
-class KnowledgeCollectionExportResponse(BaseModel):
-    """Response for POST /collections/{collection_id}/export."""
-
-    status: str
-    collection: Optional[Dict[str, Any]] = None
-    facts: List[Any]
-    total_count: int
-    exported_at: Optional[str] = None
-
-
-class KnowledgeCollectionBulkDeleteResponse(BaseModel):
-    """Response for POST /collections/{collection_id}/bulk-delete."""
-
-    status: str
-    collection_id: Optional[str] = None
-    facts_to_delete: Optional[int] = None
-    deleted_count: int
-    confirm_required: bool
-    message: Optional[str] = None
-
-
-# ---------------------------------------------------------------------------
-# knowledge_categories.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeCategoryCreateResponse(BaseModel):
-    """Response for POST /categories."""
-
-    status: str
-    category: Optional[Dict[str, Any]] = None
-    message: Optional[str] = None
-
-
-class KnowledgeCategoryTreeResponse(BaseModel):
-    """Response for GET /categories/tree."""
-
-    status: str
-    tree: List[Any]
-    total_categories: int
-
-
-class KnowledgeCategoryDetailResponse(BaseModel):
-    """Response for GET /categories/{category_id} and GET /categories/path/{path}."""
-
-    status: str
-    category: Optional[Dict[str, Any]] = None
-
-
-class KnowledgeCategoryUpdateResponse(BaseModel):
-    """Response for PUT /categories/{category_id}."""
-
-    status: str
-    category: Optional[Dict[str, Any]] = None
-    message: Optional[str] = None
-
-
-class KnowledgeCategoryDeleteResponse(BaseModel):
-    """Response for DELETE /categories/{category_id}."""
-
-    status: str
-    deleted_count: int
-    facts_reassigned: int
-    message: Optional[str] = None
-
-
-class KnowledgeCategoryChildrenResponse(BaseModel):
-    """Response for GET /categories/{category_id}/children."""
-
-    status: str
-    parent_id: Optional[str] = None
-    children: List[Any]
-    count: int
-
-
-class KnowledgeCategoryAncestorsResponse(BaseModel):
-    """Response for GET /categories/{category_id}/ancestors."""
-
-    status: str
-    category_id: Optional[str] = None
-    ancestors: List[Any]
-    depth: int
-
-
-class KnowledgeCategoryFactsResponse(BaseModel):
-    """Response for GET /categories/{category_id}/facts."""
-
-    status: str
-    category_id: Optional[str] = None
-    category_path: Optional[str] = None
-    facts: List[Any]
-    total_count: int
-    returned_count: int
-    limit: int
-    offset: int
-    has_more: bool
-    include_descendants: bool
-
-
-class KnowledgeFactAssignCategoryResponse(BaseModel):
-    """Response for POST /facts/{fact_id}/category."""
-
-    status: str
-    fact_id: Optional[str] = None
-    category_id: Optional[str] = None
-    category_path: Optional[str] = None
-    message: Optional[str] = None
-
-
-class KnowledgeCategorySearchResponse(BaseModel):
-    """Response for POST /categories/search."""
-
-    status: str
-    pattern: Optional[str] = None
-    categories: List[Any]
-    count: int
-
-
-# ---------------------------------------------------------------------------
-# validation_dashboard.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class ValidationDashboardStatusResponse(BaseModel):
-    """Response for GET /status (healthy path; unavailable path uses JSONResponse)."""
-
-    status: str
-    service: str
-    output_directory: Optional[str] = None
-    refresh_interval: Optional[int] = None
-    data_retention_days: Optional[int] = None
-    timestamp: str
-
-
-class ValidationDashboardReportResponse(BaseModel):
-    """Response for GET /report."""
-
-    status: str
-    report: Optional[Any] = None
-    timestamp: str
-
-
-class ValidationDashboardGenerateResponse(BaseModel):
-    """Response for POST /generate."""
-
-    status: str
-    message: str
-    settings: Dict[str, Any]
-    timestamp: str
-
-
-class ValidationDashboardMetricsResponse(BaseModel):
-    """Response for GET /metrics."""
-
-    status: str
-    metrics: Dict[str, Any]
-    timestamp: str
-
-
-class ValidationDashboardTrendsResponse(BaseModel):
-    """Response for GET /trends."""
-
-    status: str
-    trends: Optional[Any] = None
-    timestamp: str
-
-
-class ValidationDashboardAlertsResponse(BaseModel):
-    """Response for GET /alerts."""
-
-    status: str
-    alerts: List[Any]
-    alert_counts: Dict[str, int]
-    timestamp: str
-
-
-class ValidationDashboardRecommendationsResponse(BaseModel):
-    """Response for GET /recommendations."""
-
-    status: str
-    recommendations: List[Any]
-    recommendation_counts: Dict[str, int]
-    timestamp: str
-
-
-class ValidationJudgmentResponse(BaseModel):
-    """Response for POST /judge_workflow_step and POST /judge_agent_response."""
-
-    status: str
-    judgment: Dict[str, Any]
-    timestamp: str
-
-
-class ValidationJudgeStatusResponse(BaseModel):
-    """Response for GET /judge_status (healthy path; unavailable path uses JSONResponse)."""
-
-    status: str
-    service: str
-    available_judges: List[str]
-    judge_metrics: Dict[str, Any]
-    timestamp: str
-
-
-
-# ---------------------------------------------------------------------------
-# templates.py schemas
-# ---------------------------------------------------------------------------
-
-
-class TemplatesRootResponse(BaseModel):
-    """Response for GET / (templates root)."""
-
-    message: str
-    endpoints: Dict[str, str]
-
-
-class TemplateListResponse(BaseModel):
-    """Response for GET /templates."""
-
-    success: bool
-    templates: List[Any]
-    total: int
-
-
-class TemplateSecretsUsageResponse(BaseModel):
-    """Response for GET /templates/secrets-usage."""
-
-    success: bool
-    secrets_usage: Dict[str, Any]
-
-
-class TemplateSearchResponse(BaseModel):
-    """Response for GET /templates/search."""
-
-    success: bool
-    query: str
-    results: List[Any]
-    total: int
-
-
-class TemplateCategoriesResponse(BaseModel):
-    """Response for GET /templates/categories."""
-
-    success: bool
-    categories: List[Any]
-
-
-class TemplateStatsResponse(BaseModel):
-    """Response for GET /templates/stats."""
-
-    success: bool
-    statistics: Dict[str, Any]
-
-
-class TemplateDetailResponse(BaseModel):
-    """Response for GET /templates/{template_id}."""
-
-    success: bool
-    template: Dict[str, Any]
-
-
-class TemplatePreviewResponse(BaseModel):
-    """Response for GET /templates/{template_id}/preview."""
-
-    success: bool
-    template_id: str
-    template_name: str
-    description: str
-    estimated_duration_minutes: Optional[int] = None
-    agents_involved: List[str]
-    workflow_preview: List[str]
-    variables_used: Dict[str, Any]
-    total_steps: int
-    approval_required_steps: int
-
-
-class TemplateValidationResponse(BaseModel):
-    """Response for POST /templates/{template_id}/validate."""
-
-    success: bool
-    template_id: str
-    validation: Dict[str, Any]
-
-
-class TemplateCreateWorkflowResponse(BaseModel):
-    """Response for POST /templates/{template_id}/create-workflow.
-
-    success=False path includes validation dict; success=True includes workflow.
-    Extra fields allowed to handle both paths.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-
-
-class TemplateExecuteResponse(SuccessMessageResponse):
-    """Response for POST /templates/{template_id}/execute."""
-
-    workflow_id: str
-    template_info: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# state_tracking.py schemas
-# ---------------------------------------------------------------------------
-
-
-class StateTrackingStatusResponse(BaseModel):
-    """Response for GET /status."""
-
-    status: str
-    service: str
-    timestamp: str
-    tracking_active: bool
-    snapshot_count: Optional[Any] = None
-    change_count: Optional[Any] = None
-    latest_snapshot: Optional[Any] = None
-
-
-class StateTrackingSummaryResponse(BaseModel):
-    """Response for GET /summary."""
-
-    status: str
-    summary: Dict[str, Any]
-    timestamp: str
-
-
-class StateTrackingSnapshotResponse(BaseModel):
-    """Response for POST /snapshot."""
-
-    status: str
-    message: str
-    timestamp: str
-
-
-class StateTrackingChangeResponse(BaseModel):
-    """Response for POST /change (success path only).
-
-    JSONResponse(400) for invalid change_type bypasses response_model.
-    """
-
-    status: str
-    message: str
-    change_type: str
-    timestamp: str
-
-
-class StateTrackingMilestonesResponse(BaseModel):
-    """Response for GET /milestones."""
-
-    status: str
-    milestones: Dict[str, Any]
-    achieved_count: int
-    total_count: int
-    timestamp: str
-
-
-class StateTrackingTrendsResponse(BaseModel):
-    """Response for GET /trends/{metric} (success path only).
-
-    JSONResponse(400) for invalid metrics bypasses response_model.
-    """
-
-    status: str
-    metric: str
-    days: int
-    data_points: int
-    trend_data: List[Any]
-    timestamp: str
-
-
-class StateTrackingChangesResponse(BaseModel):
-    """Response for GET /changes."""
-
-    status: str
-    changes: List[Any]
-    total_changes: int
-    showing: int
-    timestamp: str
-
-
-class StateTrackingReportResponse(BaseModel):
-    """Response for GET /report."""
-
-    status: str
-    report: str
-    format: str
-    timestamp: str
-
-
-class StateTrackingExportResponse(BaseModel):
-    """Response for POST /export."""
-
-    status: str
-    message: str
-    filename: str
-    path: str
-    format: str
-    timestamp: str
-
-
-class StateTrackingMetricsAllResponse(BaseModel):
-    """Response for GET /metrics/all."""
-
-    status: str
-    metrics: Dict[str, Any]
-    available_metrics: List[str]
-    timestamp: str
-
-
-class StateTrackingPhaseHistoryResponse(BaseModel):
-    """Response for GET /phase-history/{phase_name}."""
-
-    status: str
-    phase_name: str
-    days: int
-    data_points: int
-    history: List[Any]
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# secrets.py schemas  (all endpoints use JSONResponse → response_model=None)
-# development_speedup.py schemas  (all endpoints use JSONResponse → response_model=None)
-# No new Pydantic models needed for those two files.
-# ---------------------------------------------------------------------------
-
-
-# ---------------------------------------------------------------------------
-# analytics_code_review.py schemas
-# ---------------------------------------------------------------------------
-
-
-class CodeReviewAnalyzeResponse(BaseModel):
-    """Response for GET /analyze (code review diff analysis).
-
-    Returns either the no-data envelope or the full analysis payload.
-    Extra fields allowed to cover both shapes.
-    """
-
-    model_config = {"extra": "allow"}
-
-    status: str
-
-
-class CodeReviewReviewByIdResponse(BaseModel):
-    """Response for GET /review/{review_id}.
-
-    Shape is the opaque result_payload dict stored in Redis — allow extra fields.
-    """
-
-    model_config = {"extra": "allow"}
-
-
-class CodeReviewFileResponse(BaseModel):
-    """Response for POST /review-file."""
-
-    status: str
-    file_path: Optional[str] = None
-    timestamp: str
-    total_comments: int
-    score: float
-    comments: List[Any]
-    summary: Dict[str, Any]
-
-
-class CodeReviewHistoryResponse(BaseModel):
-    """Response for GET /history.
-
-    Returns either no-data envelope or success+reviews shape.
-    Extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    status: str
-
-
-class CodeReviewMetricsResponse(BaseModel):
-    """Response for GET /metrics — no-data envelope."""
-
-    model_config = {"extra": "allow"}
-
-    status: str
-
-
-class CodeReviewFeedbackResponse(BaseModel):
-    """Response for POST /feedback."""
-
-    status: str
-    feedback: Dict[str, Any]
-
-
-class CodeReviewSummaryResponse(BaseModel):
-    """Response for GET /summary — no-data envelope."""
-
-    model_config = {"extra": "allow"}
-
-    status: str
-
-
-class CodeReviewPatternToggleResponse(BaseModel):
-    """Response for POST /patterns/toggle."""
-
-    status: str
-    pattern_id: str
-    enabled: bool
-
-
-class CodeReviewPatternPreferencesResponse(BaseModel):
-    """Response for GET /patterns/preferences."""
-
-    patterns: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# metrics.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class MetricsWorkflowResponse(BaseModel):
-    """Response for GET /workflow/{workflow_id}."""
-
-    success: bool
-    workflow_id: str
-    metrics: Optional[Any] = None
-
-
-class MetricsPerformanceSummaryResponse(BaseModel):
-    """Response for GET /performance/summary."""
-
-    success: bool
-    performance_summary: Optional[Any] = None
-
-
-class MetricsSystemCurrentResponse(BaseModel):
-    """Response for GET /system/current."""
-
-    success: bool
-    system_metrics: Optional[Any] = None
-
-
-class MetricsSystemHistoryResponse(BaseModel):
-    """Response for GET /system/history."""
-
-    success: bool
-    cpu_history: List[Any]
-    memory_history: List[Any]
-    time_range: Dict[str, str]
-
-
-class MetricsSystemSummaryResponse(BaseModel):
-    """Response for GET /system/summary."""
-
-    success: bool
-    resource_summary: Optional[Any] = None
-
-
-class MetricsExportResponse(BaseModel):
-    """Response for GET /export/workflow and GET /export/system."""
-
-    success: bool
-    format: str
-    data: Optional[Any] = None
-
-
-class MetricsMonitoringStartResponse(SuccessMessageResponse):
-    """Response for POST /system/monitoring/start."""
-
-    collection_interval: Optional[Any] = None
-
-
-class MetricsMonitoringStopResponse(SuccessMessageResponse):
-    """Response for POST /system/monitoring/stop."""
-
-
-class MetricsDashboardResponse(BaseModel):
-    """Response for GET /dashboard."""
-
-    success: bool
-    dashboard: Dict[str, Any]
-
-
-
-# ---------------------------------------------------------------------------
-# logs.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class LogSourcesResponse(BaseModel):
-    """Response for GET /logs/sources."""
-
-    file_logs: List[Any]
-    container_logs: List[Any]
-    total_sources: int
-
-
-class LogRecentResponse(BaseModel):
-    """Response for GET /logs/recent."""
-
-    entries: List[Any]
-    count: int
-    limit: int
-    source: str
-    error: Optional[str] = None
-
-
-class LogReadResponse(BaseModel):
-    """Response for GET /logs/read/{filename}."""
-
-    filename: str
-    lines: List[str]
-    total_lines: int
-    offset: int
-    count: int
-
-
-class LogContainerResponse(BaseModel):
-    """Response for GET /logs/container/{service}."""
-
-    service: str
-    container: str
-    lines: List[Any]
-    count: int
-    source_type: str
-
-
-class LogUnifiedResponse(BaseModel):
-    """Response for GET /logs/unified."""
-
-    logs: List[Any]
-    total_count: int
-    sources_included: List[str]
-
-
-class LogSearchResponse(BaseModel):
-    """Response for GET /logs/search."""
-
-    query: str
-    results: List[Any]
-    count: int
-    truncated: bool
-
-
-# ---------------------------------------------------------------------------
-# git_mcp.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class GitMCPOperationResponse(BaseModel):
-    """Shared response for all POST /git/mcp/* operation endpoints.
-
-    All git MCP operation endpoints return the same stable shape.
-    Extra fields (repository-specific) are allowed through.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    repository: str
-    output: str
-    timestamp: str
-    errors: Optional[str] = None
-
-
-class GitMCPInfoResponse(BaseModel):
-    """Response for GET /git/mcp/info."""
-
-    success: bool
-    repositories: List[Any]
-    repository_count: int
-    timestamp: str
-
-
-class GitMCPServiceStatusResponse(BaseModel):
-    """Response for GET /git/mcp/service_status."""
-
-    status: str
-    service: str
-    rate_limit: Dict[str, Any]
-    configuration: Dict[str, Any]
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# analytics_llm_patterns.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class LLMPatternsHealthResponse(BaseModel):
-    """Response for GET /llm-patterns/health."""
-
-    status: str
-    service: str
-    deprecated: bool
-    use_instead: str
-    features: List[str]
-    supported_categories: List[str]
-    optimization_types: List[str]
-
-
-class LLMPatternsAnalyzeResponse(BaseModel):
-    """Response for POST /llm-patterns/analyze."""
-
-    prompt_hash: str
-    category: str
-    estimated_tokens: int
-    estimated_cost: float
-    issues: List[Any]
-    recommendations: List[str]
-    cache_potential: bool
-
-
-class LLMPatternsRecordResponse(BaseModel):
-    """Response for POST /llm-patterns/record."""
-
-    recorded: bool
-    prompt_hash: str
-    category: str
-    cost: float
-    cache_count: int
-
-
-class LLMPatternsStatsResponse(BaseModel):
-    """Response for GET /llm-patterns/stats.
-
-    Shape includes totals and by_date/by_model/by_category which are dynamic;
-    extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    period_days: int
-    total_requests: int
-    total_cost: float
-    successful_requests: int
-
-
-class LLMPatternsCacheOpportunitiesResponse(BaseModel):
-    """Response for GET /llm-patterns/cache-opportunities."""
-
-    opportunities: List[Any]
-    count: int
-    min_occurrences: int
-
-
-class LLMPatternsRecommendationsResponse(BaseModel):
-    """Response for GET /llm-patterns/recommendations."""
-
-    recommendations: List[Any]
-
-
-class LLMPatternsModelComparisonResponse(BaseModel):
-    """Response for GET /llm-patterns/model-comparison."""
-
-    models: List[Any]
-    period_days: int
-
-
-class LLMPatternsCategoryDistributionResponse(BaseModel):
-    """Response for GET /llm-patterns/category-distribution."""
-
-    categories: List[Any]
-    total_count: int
-    total_cost: float
-
-
-class LLMPatternsCostBreakdownResponse(BaseModel):
-    """Response for GET /llm-patterns/cost-breakdown."""
-
-    period_days: int
-    total_cost: float
-    total_requests: int
-    avg_cost_per_request: float
-    by_model: List[Any]
-    by_category: List[Any]
-    daily_trend: List[Any]
-
-
-# ---------------------------------------------------------------------------
-# voice.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class VoiceListenResponse(BaseModel):
-    """Response for POST /voice/listen (success path)."""
-
-    message: str
-    text: str
-
-
-class VoiceSpeakResponse(BaseModel):
-    """Response for POST /voice/speak (success path)."""
-
-    message: str
-
-
-class VoiceDeleteResponse(BaseModel):
-    """Response for DELETE /voice/voices/{voice_id} (success path)."""
-
-    deleted: str
-
-
-class VoiceTranscribeResponse(BaseModel):
-    """Response for POST /voice/transcribe."""
-
-    text: str
-    language: str
-    confidence: float
-
-
-# ---------------------------------------------------------------------------
-# skills_governance.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class SkillsGapResponse(BaseModel):
-    """Response for POST /skills-governance/gaps.
-
-    Both success and failure paths share this shape.
-    """
-
-    success: bool
-    errors: Optional[List[str]] = None
-    draft: Optional[Any] = None
-    draft_id: Optional[str] = None
-    name: Optional[str] = None
-    tools_found: Optional[List[str]] = None
-
-
-class SkillsDraftListItem(BaseModel):
-    """Single draft entry in the list response."""
-
-    id: str
-    name: str
-    version: Optional[Any] = None
-    gap_reason: Optional[str] = None
-    created_at: Optional[Any] = None
-    trust_level: Optional[Any] = None
-
-
-class SkillsDraftTestResponse(BaseModel):
-    """Response for POST /skills-governance/drafts/{skill_id}/test."""
-
-    valid: bool
-    errors: List[Any]
-    tools_found: List[str]
-
-
-class SkillsDraftPromoteResponse(BaseModel):
-    """Response for POST /skills-governance/drafts/{skill_id}/promote."""
-
-    promoted: bool
-    path: str
-    name: str
-
-
-class SkillsApprovalItem(BaseModel):
-    """Single approval entry in the list response."""
-
-    id: str
-    skill_id: str
-    requested_by: Optional[str] = None
-    requested_at: Optional[Any] = None
-    reason: Optional[str] = None
-    status: str
-
-
-class SkillsApprovalDecisionResponse(BaseModel):
-    """Response for POST /skills-governance/approvals/{approval_id}."""
-
-    approval_id: str
-    status: str
-
-
-class SkillsGovernanceConfigResponse(BaseModel):
-    """Response for GET /skills-governance/ (may return default dict or full config).
-
-    Extra fields allowed to cover the default-dict path.
-    """
-
-    model_config = {"extra": "allow"}
-
-    mode: Any
-
-
-class SkillsGovernanceUpdateResponse(BaseModel):
-    """Response for PUT /skills-governance/."""
-
-    mode: Any
-
-
-
-# ---------------------------------------------------------------------------
-# mcp_registry.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class MCPRegistryToolsResponse(BaseModel):
-    """Response for GET /tools — aggregated tool list from all bridges."""
-
-    status: str
-    total_tools: int
-    total_bridges: int
-    healthy_bridges: int
-    tools: List[Any]
-    last_updated: str
-    cached: bool
-
-
-class MCPRegistryBridgesResponse(BaseModel):
-    """Response for GET /bridges — bridge health and metadata."""
-
-    status: str
-    total_bridges: int
-    healthy_bridges: int
-    bridges: List[Any]
-    last_checked: str
-    cached: bool
-
-
-class MCPRegistryCacheInvalidateResponse(BaseModel):
-    """Response for POST /cache/invalidate."""
-
-    status: str
-    message: str
-    timestamp: str
-    cache_stats: Dict[str, Any]
-
-
-class MCPRegistryCacheStatsResponse(BaseModel):
-    """Response for GET /cache/stats."""
-
-    status: str
-    cache: Dict[str, Any]
-    timestamp: str
-
-
-class MCPRegistryToolDetailResponse(BaseModel):
-    """Response for GET /tools/{bridge_name}/{tool_name}."""
-
-    status: str
-    tool: Dict[str, Any]
-
-
-class MCPRegistryHealthResponse(BaseModel):
-    """Response for GET /health."""
-
-    status: str
-    total_bridges: int
-    healthy_bridges: int
-    checks: List[Any]
-    cache_stats: Dict[str, Any]
-    timestamp: str
-
-
-class MCPRegistryStatsResponse(BaseModel):
-    """Response for GET /stats."""
-
-    status: str
-    overview: Dict[str, Any]
-    tools_by_bridge: Dict[str, Any]
-    bridge_health: Dict[str, Any]
-    available_features: List[str]
-    cache: Dict[str, Any]
-    timestamp: str
-
-
-class MCPRegistryInfoResponse(BaseModel):
-    """Response for GET / — registry info and architecture overview."""
-
-    name: str
-    version: str
-    description: str
-    architecture: Dict[str, Any]
-    endpoints: Dict[str, str]
-    performance: Dict[str, Any]
-    note: str
-
-
-# ---------------------------------------------------------------------------
-# feature_flags.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class FeatureFlagStatusResponse(BaseModel):
-    """Response for GET /feature-flags/status."""
-
-    success: bool
-    data: Optional[Any] = None
-
-
-class FeatureFlagEnforcementModeResponse(SuccessDataResponse):
-    """Response for PUT /feature-flags/enforcement-mode."""
-
-
-class FeatureFlagEndpointSetResponse(SuccessDataResponse):
-    """Response for PUT /feature-flags/endpoint/{endpoint:path}."""
-
-
-class FeatureFlagEndpointRemoveResponse(SuccessDataResponse):
-    """Response for DELETE /feature-flags/endpoint/{endpoint:path}."""
-
-
-class AccessControlMetricsResponse(BaseModel):
-    """Response for GET /access-control/metrics."""
-
-    success: bool
-    data: Optional[Dict[str, Any]] = None
-
-
-class AccessControlEndpointMetricsResponse(BaseModel):
-    """Response for GET /access-control/endpoint/{endpoint:path}."""
-
-    success: bool
-    data: Optional[Any] = None
-
-
-class AccessControlUserMetricsResponse(BaseModel):
-    """Response for GET /access-control/user/{username}."""
-
-    success: bool
-    data: Optional[Any] = None
-
-
-class AccessControlCleanupResponse(SuccessMessageResponse):
-    """Response for POST /access-control/cleanup."""
-
-
-# ---------------------------------------------------------------------------
-# http_client_mcp.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class HTTPClientMCPStatusResponse(BaseModel):
-    """Response for GET /mcp/status — HTTP client service status."""
-
-    status: str
-    service: str
-    rate_limit: Dict[str, Any]
-    configuration: Dict[str, Any]
-    timestamp: str
-
-
-class HTTPRequestResultResponse(BaseModel):
-    """Response for POST /mcp/get|post|put|patch|delete|head.
-
-    Shape is the standardised HTTP response dict from _build_http_response().
-    Extra fields allowed to handle dynamic headers dict.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    status_code: int
-    url: str
-    method: str
-    is_json: bool
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# database_mcp.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class DatabaseQueryResponse(BaseModel):
-    """Response for POST /mcp/query."""
-
-    success: bool
-    database: str
-    query: str
-    row_count: int
-    columns: List[str]
-    results: List[Any]
-    timestamp: str
-
-
-class DatabaseExecuteResponse(BaseModel):
-    """Response for POST /mcp/execute."""
-
-    success: bool
-    database: str
-    statement: str
-    rows_affected: int
-    timestamp: str
-
-
-class DatabaseListTablesResponse(BaseModel):
-    """Response for POST /mcp/list_tables."""
-
-    success: bool
-    database: str
-    table_count: int
-    tables: List[Any]
-    timestamp: str
-
-
-class DatabaseDescribeSchemaResponse(BaseModel):
-    """Response for POST /mcp/describe_schema."""
-
-    success: bool
-    database: str
-    table_count: int
-    schemas: Dict[str, Any]
-    timestamp: str
-
-
-class DatabaseListDatabasesResponse(BaseModel):
-    """Response for GET /mcp/list_databases."""
-
-    success: bool
-    database_count: int
-    databases: List[Any]
-    timestamp: str
-
-
-class DatabaseStatisticsResponse(BaseModel):
-    """Response for POST /mcp/statistics."""
-
-    success: bool
-    database: str
-    statistics: Dict[str, Any]
-    timestamp: str
-
-
-class DatabaseMCPStatusResponse(BaseModel):
-    """Response for GET /mcp/status — database service status."""
-
-    status: str
-    service: str
-    rate_limit: Dict[str, Any]
-    configuration: Dict[str, Any]
-    database_availability: Dict[str, Any]
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# knowledge_search_aggregator.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeUnifiedSearchResponse(BaseModel):
-    """Response for POST /unified/search."""
-
-    success: bool
-    query: str
-    facts: List[Any]
-    related_facts: List[Any]
-    documentation: List[Any]
-    sources_searched: List[str]
-    total_results: int
-
-
-class KnowledgeUnifiedStatsResponse(BaseModel):
-    """Response for GET /unified/stats.
-
-    Sections are populated dynamically — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    knowledge_base: Dict[str, Any]
-    relations: Dict[str, Any]
-    documentation: Dict[str, Any]
-
-
-class KnowledgeUnifiedContextResponse(BaseModel):
-    """Response for POST /unified/context."""
-
-    success: bool
-    context: str
-    context_length: int
-    citations: List[Any]
-    sources_used: List[Any]
-
-
-class KnowledgeDocumentationSearchResponse(BaseModel):
-    """Response for GET /unified/documentation/search."""
-
-    success: bool
-    query: Optional[str] = None
-    results: List[Any]
-    total_results: Optional[int] = None
-    message: Optional[str] = None
-
-
-class KnowledgeDocumentationStatsResponse(BaseModel):
-    """Response for GET /unified/documentation/stats."""
-
-    success: bool
-    indexed: Optional[bool] = None
-    message: Optional[str] = None
-    how_to_index: Optional[str] = None
-    collection_name: Optional[str] = None
-    document_count: Optional[int] = None
-
-
-class KnowledgeUnifiedGraphResponse(BaseModel):
-    """Response for POST /unified/graph and GET /unified/graph."""
-
-    success: bool
-    data: Dict[str, Any]
-    stats: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# knowledge_relations.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeRelationResultResponse(BaseModel):
-    """Generic response for relation create/delete/traverse/hybrid-search/stats endpoints.
-
-    KB methods return opaque success dicts — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-
-
-class KnowledgeRelationTypesResponse(BaseModel):
-    """Response for GET /types."""
-
-    success: bool
-    relation_types: List[Dict[str, Any]]
-
-
-# ---------------------------------------------------------------------------
-# knowledge_collaboration.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeScopedFactsResponse(BaseModel):
-    """Response for GET /knowledge/collaboration/facts and /facts/organization/{id} and /facts/group/{id}."""
-
-    facts: List[Any]
-    count: int
-    total: Optional[int] = None
-
-
-class KnowledgeShareResponse(BaseModel):
-    """Response for POST /knowledge/collaboration/facts/{id}/share."""
-
-    success: bool
-    fact_id: str
-    visibility: Optional[str] = None
-    shared_with: List[Any]
-    group_ids: List[Any]
-
-
-class KnowledgeUnshareResponse(BaseModel):
-    """Response for DELETE /knowledge/collaboration/facts/{id}/share/{entity_id}."""
-
-    success: bool
-    fact_id: str
-    visibility: Optional[str] = None
-    shared_with: List[Any]
-    group_ids: List[Any]
-
-
-class KnowledgePermissionsUpdateResponse(BaseModel):
-    """Response for PUT /knowledge/collaboration/facts/{id}/permissions."""
-
-    success: bool
-    fact_id: str
-    visibility: Optional[str] = None
-    organization_id: Optional[str] = None
-    group_ids: List[Any]
-
-
-class KnowledgeAccessInfoResponse(BaseModel):
-    """Response for GET /knowledge/collaboration/facts/{id}/access."""
-
-    fact_id: str
-    owner_id: Optional[str] = None
-    visibility: Optional[str] = None
-    organization_id: Optional[str] = None
-    group_ids: List[Any]
-    shared_with: List[Any]
-    can_edit: bool
-    can_share: bool
-    can_delete: bool
-    has_access: bool
-
-
-# ---------------------------------------------------------------------------
-# knowledge_audit.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeAuditEventsResponse(BaseModel):
-    """Response for GET /knowledge/audit/user-activity, /fact/{id}/access-log, /organization/audit-log."""
-
-    events: List[Any]
-    count: int
-    user_id: Optional[str] = None
-    fact_id: Optional[str] = None
-    organization_id: Optional[str] = None
-
-
-class KnowledgePermissionChangesResponse(BaseModel):
-    """Response for GET /knowledge/audit/permission-changes."""
-
-    events: List[Any]
-    count: int
-
-
-class KnowledgeComplianceReportResponse(BaseModel):
-    """Response for POST /knowledge/audit/compliance-report and GET /compliance-summary.
-
-    Shape is defined by AuditLog.generate_compliance_report() — allow extra fields.
-    """
-
-    model_config = {"extra": "allow"}
-
-    total_events: int
-
-
-# ---------------------------------------------------------------------------
-# knowledge_verification.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeVerificationPendingResponse(BaseModel):
-    """Response for GET /verification/pending."""
-
-    status: str
-    pending: List[Any]
-    total: int
-    limit: int
-    offset: int
-    has_more: bool
-
-
-class KnowledgeVerificationApproveResponse(BaseModel):
-    """Response for POST /verification/{fact_id}/approve."""
-
-    status: str
-    fact_id: str
-    verified_by: Optional[str] = None
-    verified_at: Optional[str] = None
-    message: Optional[str] = None
-
-
-class KnowledgeVerificationRejectResponse(BaseModel):
-    """Response for POST /verification/{fact_id}/reject."""
-
-    status: str
-    fact_id: str
-    deleted: bool
-    message: Optional[str] = None
-
-
-class KnowledgeVerificationConfigResponse(BaseModel):
-    """Response for GET/PUT /verification/config."""
-
-    status: str
-    config: Dict[str, Any]
-    message: Optional[str] = None
-
-
-# ---------------------------------------------------------------------------
-# knowledge_suggestions.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeSuggestionsTagsResponse(BaseModel):
-    """Response for POST /suggestions/tags.
-
-    KB method returns opaque success dict — allow extra fields.
-    """
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    suggestions: Optional[List[Any]] = None
-    similar_docs_analyzed: Optional[int] = None
-
-
-class KnowledgeSuggestionsCategoriesResponse(BaseModel):
-    """Response for POST /suggestions/categories."""
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    suggestions: Optional[List[Any]] = None
-    similar_docs_analyzed: Optional[int] = None
-
-
-class KnowledgeSuggestionsAllResponse(BaseModel):
-    """Response for POST /suggestions/all."""
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-
-
-class KnowledgeSuggestionsContextResponse(BaseModel):
-    """Response for POST /suggestions/context."""
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-    suggestions: Optional[List[Any]] = None
-    total_candidates: Optional[int] = None
-
-
-class KnowledgeAutoApplySuggestionsResponse(BaseModel):
-    """Response for POST /facts/{fact_id}/auto-apply."""
-
-    model_config = {"extra": "allow"}
-
-    success: bool
-
-
-# ---------------------------------------------------------------------------
-# knowledge_ownership.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeShareFactResponse(BaseModel):
-    """Response for POST /api/knowledge/facts/{fact_id}/share."""
-
-    success: bool
-    fact_id: str
-    shared_with: List[Any]
-    visibility: Optional[str] = None
-
-
-class KnowledgeUnshareFactResponse(BaseModel):
-    """Response for DELETE /api/knowledge/facts/{fact_id}/share/{user_id_to_remove}."""
-
-    success: bool
-    fact_id: str
-    shared_with: List[Any]
-    visibility: Optional[str] = None
-
-
-class KnowledgeUpdateVisibilityResponse(BaseModel):
-    """Response for PUT /api/knowledge/facts/{fact_id}/visibility."""
-
-    success: bool
-    fact_id: str
-    visibility: Optional[str] = None
-
-
-class KnowledgeMyFactsResponse(BaseModel):
-    """Response for GET /api/knowledge/facts/mine."""
-
-    success: bool
-    user_id: str
-    owned_count: int
-    shared_count: int
-    facts: List[Any]
-    total_returned: int
-
-
-class KnowledgeSharedWithMeResponse(BaseModel):
-    """Response for GET /api/knowledge/facts/shared-with-me."""
-
-    success: bool
-    user_id: str
-    facts: List[Any]
-    total_returned: int
-
-
-# ---------------------------------------------------------------------------
-# knowledge_grounding.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeGroundResponseResponse(BaseModel):
-    """Response for POST /api/ground-response."""
-
-    status: str
-    data: Dict[str, Any]
-
-
-class KnowledgeVerifyClaimResponse(BaseModel):
-    """Response for POST /api/verify-claim."""
-
-    status: str
-    claim_text: str
-    kb_status: str
-    confidence: float
-    evidence: Optional[List[Any]] = None
-    verification_method: Optional[str] = None
-    kb_source: Optional[str] = None
-
-
-class KnowledgeConflictsListResponse(BaseModel):
-    """Response for GET /api/kb-conflicts."""
-
-    status: str
-    conflicts: List[Any]
-    total: int
-    limit: int
-    offset: int
-    has_more: bool
-
-
-class KnowledgeResolveConflictResponse(BaseModel):
-    """Response for POST /api/kb-conflicts/{conflict_id}/resolve."""
-
-    status: str
-    data: Dict[str, Any]
-
-
-class KnowledgeGroundingStatsResponse(BaseModel):
-    """Response for GET /api/kb-stats.
-
-    Stats shape is dynamic — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    status: str
-    period: str
-
-
-# ---------------------------------------------------------------------------
-# knowledge_organization.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeOrganizationPolicyResponse(BaseModel):
-    """Response for GET/PUT /knowledge/organization/policy.
-
-    Returns OrganizationKnowledgePolicy fields — allow extra fields.
-    """
-
-    model_config = {"extra": "allow"}
-
-    default_visibility: Any
-    allow_user_private: bool
-    allow_user_shared: bool
-    allow_user_organization: bool
-    require_approval_for_system: bool
-    retention_days: Optional[int] = None
-
-
-class KnowledgeOrganizationStatsResponse(BaseModel):
-    """Response for GET /knowledge/organization/stats."""
-
-    organization_id: str
-    total_facts: int
-    by_visibility: Dict[str, Any]
-    by_source: Dict[str, Any]
-    total_size_bytes: int
-    user_count: int
-    team_count: int
-    top_contributors: List[Any]
-
-
-class KnowledgeOrganizationCleanupResponse(BaseModel):
-    """Response for DELETE /knowledge/organization/cleanup."""
-
-    success: bool
-    organization_id: str
-    retention_days: int
-    deleted_count: int
-    cutoff_date: str
-
-
-# ---------------------------------------------------------------------------
-# knowledge_search_scoped.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeScopedSearchResponse(BaseModel):
-    """Response for POST /knowledge/search/scoped and POST /knowledge/search/rag/scoped."""
-
-    model_config = {"extra": "allow"}
-
-    results: Optional[List[Any]] = None
-    total_results: Optional[int] = None
-    query: Optional[str] = None
-    mode: Optional[str] = None
-    user_id: Optional[str] = None
-    filtered_by_permissions: Optional[bool] = None
-
-
-class KnowledgeAccessibleScopesResponse(BaseModel):
-    """Response for GET /knowledge/search/accessible-scopes."""
-
-    user_id: Optional[str] = None
-    organization_id: Optional[str] = None
-    group_count: int
-    accessible_scopes: List[Any]
-
-
-# ---------------------------------------------------------------------------
-# knowledge_debug.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeFreshStatsResponse(BaseModel):
-    """Response for GET /fresh_stats.
-
-    Shape varies between success and error paths — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    source: str
-    total_facts: int
-    status: str
-
-
-class KnowledgeDebugRedisResponse(BaseModel):
-    """Response for GET /debug_redis.
-
-    Shape varies between success and failure — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    redis_connection: str
-
-
-class KnowledgeRebuildIndexResponse(BaseModel):
-    """Response for POST /rebuild_index.
-
-    Shape varies depending on result.get('status') — extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-    operation: str
-    success: bool
-
-
-# ---------------------------------------------------------------------------
-# knowledge_boards.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeBoardsListResponse(BaseModel):
-    """Response for GET /boards."""
-
-    boards: List[Any]
-    total: int
-
-
-class KnowledgeBoardCreateResponse(BaseModel):
-    """Response for POST /boards."""
-
-    board_id: str
-    name: str
-    created: bool
-
-
-class KnowledgeBoardDeleteResponse(BaseModel):
-    """Response for DELETE /boards/{board_id}."""
-
-    board_id: str
-    deleted: bool
-
-
-# ---------------------------------------------------------------------------
-# knowledge_cognition.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeCognitionStatusResponse(BaseModel):
-    """Response for GET /cognition-store/status."""
-
-    collections: List[Any]
-    total_seeded_collections: int
-
-
-class KnowledgeCognitionSeedResponse(BaseModel):
-    """Response for POST /cognition-store/seed."""
-
-    status: str
-    manifest: str
-
-
-# ---------------------------------------------------------------------------
-# knowledge_sync_queue.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeSyncQueueResponse(BaseModel):
-    """Response for GET /knowledge/sync-queue."""
-
-    pending: List[Any]
-    failed: List[Any]
-    counts: Dict[str, Any]
-    limit: int
-    offset: int
-
-
-class KnowledgeSyncQueuePruneResponse(BaseModel):
-    """Response for POST /knowledge/sync-queue/prune."""
-
-    pruned: int
-
-
-
-
-# ---------------------------------------------------------------------------
-# workflow_export.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class WorkflowExportResponse(BaseModel):
-    """Response for GET /export/{workflow_id}."""
-
-    success: bool
-    export: Dict[str, Any]
-
-
-class WorkflowValidateImportResponse(BaseModel):
-    """Response for POST /validate."""
-
-    success: bool
-    valid: bool
-    issues: List[Any]
-
-
-class WorkflowImportResponse(BaseModel):
-    """Response for POST /import and POST /share/{share_id}/clone."""
-
-    success: bool
-    workflow_id: str
-
-
-class WorkflowShareResponse(BaseModel):
-    """Response for POST /share."""
-
-    success: bool
-    share_id: str
-
-
-class WorkflowUnshareResponse(BaseModel):
-    """Response for DELETE /share/{share_id}."""
-
-    success: bool
-    share_id: str
-
-
-class WorkflowListSharesResponse(BaseModel):
-    """Response for GET /shares."""
-
-    success: bool
-    shares: List[Any]
-    total_count: int
-
-
-# ---------------------------------------------------------------------------
-# rum.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class RUMConfigResponse(BaseModel):
-    """Response for POST /rum/config."""
-
-    status: str
-    message: str
-    config: Dict[str, Any]
-
-
-class RUMEventResponse(BaseModel):
-    """Response for POST /rum/event (both success and disabled/error paths).
-
-    'session_event_count' is absent on the disabled/error path — Optional.
-    """
-
-    status: str
-    message: str
-    session_event_count: Optional[int] = None
-
-
-class RUMDisableResponse(BaseModel):
-    """Response for POST /rum/disable."""
-
-    status: str
-    message: str
-
-
-class RUMClearResponse(BaseModel):
-    """Response for POST /rum/clear."""
-
-    status: str
-    message: str
-    events_cleared: int
-    sessions_cleared: int
-
-
-class RUMStatusResponse(BaseModel):
-    """Response for GET /rum/status."""
-
-    status: str
-    rum_status: Dict[str, Any]
-
-
-class RUMMetricsResponse(BaseModel):
-    """Response for POST /rum/metrics (both success and error paths).
-
-    'metrics_recorded' is absent on the error path — Optional.
-    """
-
-    status: str
-    message: str
-    session_id: str
-    metrics_recorded: Optional[int] = None
-
-
-# ---------------------------------------------------------------------------
-# registry.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class RegistryEndpointsResponse(BaseModel):
-    """Response for GET /endpoints."""
-
-    endpoints: List[Any]
-    total: int
-
-
-class RegistryRouterDetailResponse(BaseModel):
-    """Response for GET /router/{router_name}.
-
-    Includes a possible 'error' key when not found; extra fields allowed.
-    """
-
-    model_config = {"extra": "allow"}
-
-
-class RegistryTagsResponse(BaseModel):
-    """Response for GET /tags."""
-
-    tags: List[str]
-
-
-class RegistryTagRoutersResponse(BaseModel):
-    """Response for GET /tags/{tag}."""
-
-    tag: str
-    routers: List[Any]
-    count: int
-
-
-class RegistryValidateResponse(BaseModel):
-    """Response for GET /validate."""
-
-    valid: bool
-    errors: Dict[str, Any]
-
-
-class RegistryHealthResponse(BaseModel):
-    """Response for GET /health."""
-
-    status: str
-    total_routers: int
-    enabled_routers: int
-    disabled_routers: int
-
-
-# ---------------------------------------------------------------------------
-# elevation.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class ElevationRequestResponse(SuccessMessageResponse):
-    """Response for POST /elevation/request."""
-
-    request_id: str
-
-
-class ElevationAuthorizeResponse(SuccessMessageResponse):
-    """Response for POST /elevation/authorize."""
-
-    session_token: str
-    expires_in: int
-
-
-class ElevationStatusResponse(BaseModel):
-    """Response for GET /elevation/status/{request_id}."""
-
-    success: bool
-    request_id: str
-    status: str
-    operation: str
-    timestamp: Any  # datetime object serialised as string by FastAPI
-
-
-class ElevationExecuteResponse(BaseModel):
-    """Response for POST /elevation/execute/{session_token}."""
-
-    success: bool
-    output: str
-    error: str
-    return_code: int
-
-
-class ElevationPendingResponse(BaseModel):
-    """Response for GET /elevation/pending."""
-
-    success: bool
-    pending_requests: Dict[str, Any]
-    count: int
-
-
-class ElevationHealthResponse(BaseModel):
-    """Response for GET /elevation/health."""
-
-    status: str
-    service: str
-    active_sessions: int
-    pending_requests: int
-    timestamp: str
-
-
-# ---------------------------------------------------------------------------
-# usage.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class UsageSummaryPeriod(BaseModel):
-    days: int
-    start: str
-    end: str
-
-
-class UsageSummaryTokens(BaseModel):
-    input: int
-    output: int
-    total: int
-
-
-class UsageSummaryResponse(BaseModel):
-    """Response for GET /usage/summary."""
-
-    period: UsageSummaryPeriod
-    tokens: UsageSummaryTokens
-    cost_usd: float
-    requests: int
-    daily_costs: Dict[str, Any]
-    by_model: Dict[str, Any]
-    active_users: int
-
-
-class UsageByUserAllResponse(BaseModel):
-    """Response for GET /usage/by-user."""
-
-    timestamp: str
-    users: List[Any]
-    total_users: int
-
 
 class UsageRecordResponse(BaseModel):
     """Response for POST /usage/record."""
@@ -3267,450 +84,3 @@ class UsageRecordResponse(BaseModel):
 # ---------------------------------------------------------------------------
 # structured_thinking_mcp.py schemas  (Issue #5317)
 # ---------------------------------------------------------------------------
-
-
-class StructuredThinkingClearResponse(SuccessMessageResponse):
-    """Response for POST /mcp/clear_history."""
-
-    session_id: str
-    thoughts_cleared: int
-
-
-class StructuredThinkingSessionsResponse(BaseModel):
-    """Response for GET /sessions."""
-
-    session_count: int
-    sessions: List[Any]
-
-
-# ---------------------------------------------------------------------------
-# advanced_control.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class AdvancedControlStreamingTerminateResponse(BaseModel):
-    """Response for DELETE /streaming/{session_id}."""
-
-    success: bool
-    session_id: str
-
-
-class AdvancedControlStreamingSessionsResponse(BaseModel):
-    """Response for GET /streaming/sessions."""
-
-    sessions: List[Any]
-    count: int
-
-
-class AdvancedControlTakeoverRequestResponse(BaseModel):
-    """Response for POST /takeover/request."""
-
-    success: bool
-    request_id: str
-
-
-class AdvancedControlTakeoverApproveResponse(BaseModel):
-    """Response for POST /takeover/{request_id}/approve."""
-
-    success: bool
-    session_id: str
-
-
-class AdvancedControlTakeoverActionResponse(BaseModel):
-    """Response for POST /takeover/sessions/{session_id}/action."""
-
-    success: bool
-    result: Any
-
-
-class AdvancedControlTakeoverSessionStatusResponse(BaseModel):
-    """Response for pause/resume/complete takeover session endpoints."""
-
-    success: bool
-    session_id: str
-    status: str
-
-
-class AdvancedControlPendingTakeoversResponse(BaseModel):
-    """Response for GET /takeover/pending."""
-
-    pending_requests: List[Any]
-    count: int
-
-
-class AdvancedControlActiveTakeoversResponse(BaseModel):
-    """Response for GET /takeover/active."""
-
-    active_sessions: List[Any]
-    count: int
-
-
-class AdvancedControlEmergencyStopResponse(SuccessMessageResponse):
-    """Response for POST /system/emergency-stop."""
-
-    takeover_request_id: str
-
-
-class AdvancedControlSystemHealthResponse(BaseModel):
-    """Response for GET /system/health."""
-
-    status: str
-    desktop_streaming_available: bool
-    novnc_available: bool
-    active_streaming_sessions: int
-    pending_takeovers: int
-    active_takeovers: int
-    paused_tasks: int
-
-
-class AdvancedControlInfoResponse(BaseModel):
-    """Response for GET /."""
-
-    name: str
-    version: str
-    features: List[str]
-    endpoints: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# analytics.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class AnalyticsTrackEventResponse(BaseModel):
-    """Response for POST /events/track."""
-
-    status: str
-    event_id: str
-    broadcast_count: int
-
-
-class AnalyticsCollectionStartResponse(BaseModel):
-    """Response for POST /collection/start."""
-
-    status: str
-    message: str
-    session_id: str
-    metrics_collection: bool
-
-
-class AnalyticsCollectionStopResponse(BaseModel):
-    """Response for POST /collection/stop."""
-
-    status: str
-    message: str
-    session_duration: str
-
-
-class AnalyticsDashboardAnalyzeResponse(BaseModel):
-    """Response for POST /dashboard/overview/analyze."""
-
-    task_id: str
-    status: str
-
-
-class AnalyticsClearStuckTasksResponse(BaseModel):
-    """Response for POST /dashboard/overview/tasks/clear-stuck."""
-
-    cleared_count: int
-    message: str
-
-
-# ---------------------------------------------------------------------------
-# npu_workers.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class NPUStatusResponse(BaseModel):
-    """Response for GET /npu/status."""
-
-    status: str
-    total_workers: int
-    online_workers: int
-    offline_workers: int
-    total_capacity: int
-    current_load: int
-    utilization_percent: float
-    load_balancing_strategy: str
-
-
-class NPUWorkerUnpairResponse(SuccessMessageResponse):
-    """Response for POST /npu/workers/{worker_id}/unpair."""
-
-    worker_id: str
-
-
-class NPUWorkerRepairResponse(SuccessMessageResponse):
-    """Response for POST /npu/workers/{worker_id}/repair."""
-
-    old_worker_id: str
-    new_worker_id: str
-    config: Dict[str, Any]
-    server_timestamp: str
-
-
-class NPUWorkerPairResponse(SuccessMessageResponse):
-    """Response for POST /npu/workers/pair."""
-
-    worker_id: str
-    url: str
-    platform: str
-    device_info: Dict[str, Any]
-    paired_at: str
-
-
-class NPUWorkerHeartbeatResponse(BaseModel):
-    """Response for POST /npu/workers/heartbeat."""
-
-    acknowledged: bool
-    worker_id: str
-    server_timestamp: str
-    message: str
-
-
-class NPUPoolWorkersResponse(BaseModel):
-    """Response for GET /npu/pool/workers."""
-
-    workers: List[Any]
-
-
-class NPUPoolReloadResponse(SuccessMessageResponse):
-    """Response for POST /npu/pool/reload."""
-
-    workers_loaded: int
-
-
-
-
-# ---------------------------------------------------------------------------
-# wake_word.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class WakeWordListResponse(BaseModel):
-    """Response for GET /words."""
-
-    wake_words: List[str]
-    total: int
-
-
-class WakeWordMutateResponse(SuccessMessageResponse):
-    """Response for POST /words and DELETE /words/{wake_word}."""
-
-    wake_words: List[str]
-
-
-class WakeWordConfigUpdateResponse(SuccessMessageResponse):
-    """Response for PUT /config."""
-
-    config: Dict[str, Any]
-
-
-class WakeWordStatsResetResponse(SuccessMessageResponse):
-    """Response for POST /stats/reset."""
-
-    stats: Dict[str, Any]
-
-
-class WakeWordFeedbackResponse(SuccessMessageResponse):
-    """Response for POST /feedback."""
-
-    stats: Dict[str, Any]
-
-
-class WakeWordToggleResponse(SuccessMessageResponse):
-    """Response for POST /enable and POST /disable."""
-
-    config: Dict[str, Any]
-
-
-class WakeWordListeningToggleResponse(SuccessMessageResponse):
-    """Response for POST /listening/start and POST /listening/stop."""
-
-    status: Dict[str, Any]
-
-
-# ---------------------------------------------------------------------------
-# analytics_cost.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class CostModelEntry(BaseModel):
-    model: str
-    cost_usd: float
-    input_tokens: int
-    output_tokens: int
-    call_count: int
-    avg_cost_per_call: float
-
-
-class CostByModelResponse(BaseModel):
-    """Response for GET /cost/by-model."""
-
-    timestamp: str
-    models: List[CostModelEntry]
-    total_models: int
-
-
-class CostForecastPeriod(BaseModel):
-    start: str
-    days: int
-
-
-class CostForecastBaseline(BaseModel):
-    avg_daily_cost: float
-    trend: str
-    growth_rate_percent: float
-
-
-class CostForecastValues(BaseModel):
-    total_estimated_usd: float
-    daily_estimates: Dict[str, Any]
-
-
-class CostForecastResponse(BaseModel):
-    """Response for GET /cost/forecast."""
-
-    forecast_period: CostForecastPeriod
-    baseline: CostForecastBaseline
-    forecast: CostForecastValues
-    confidence: str
-
-
-class RecentUsageResponse(BaseModel):
-    """Response for GET /cost/usage/recent."""
-
-    count: int
-    records: List[Any]
-
-
-class ModelPricingEntry(BaseModel):
-    model: str
-    provider: str
-    input_price_per_1m: float
-    output_price_per_1m: float
-    is_free: bool
-
-
-class ModelPricingResponse(BaseModel):
-    """Response for GET /cost/pricing."""
-
-    pricing_date: str
-    currency: str
-    models: List[ModelPricingEntry]
-    total_models: int
-
-
-class CostEstimateResponse(BaseModel):
-    """Response for GET /cost/estimate."""
-
-    model: str
-    input_tokens: int
-    output_tokens: int
-    estimated_cost_usd: float
-    total_tokens: int
-
-
-class BudgetAlertCreateResponse(BaseModel):
-    """Response for POST /cost/budget-alert."""
-
-    status: str
-    alert: Dict[str, Any]
-
-
-class BudgetAlertsListResponse(BaseModel):
-    """Response for GET /cost/budget-alerts."""
-
-    alerts: List[Any]
-    count: int
-
-
-class BudgetStatusResponse(BaseModel):
-    """Response for GET /cost/budget-status."""
-
-    timestamp: str
-    current_costs: Dict[str, Any]
-    budget_statuses: List[Any]
-
-
-class AllAgentCostsResponse(BaseModel):
-    """Response for GET /cost/by-agent."""
-
-    timestamp: str
-    agents: List[Any]
-    total_agents: int
-
-
-# ---------------------------------------------------------------------------
-# files.py schemas  (Issue #5317)
-# ---------------------------------------------------------------------------
-
-
-class FileSandboxViewResponse(BaseModel):
-    """Response for GET /files/view/{path}."""
-
-    file_info: Any
-    content: Optional[str] = None
-    is_text: bool
-
-
-class FileSandboxRenameResponse(BaseModel):
-    """Response for POST /files/rename."""
-
-    message: str
-    item_info: Any
-
-
-class FileSandboxPreviewResponse(BaseModel):
-    """Response for GET /files/preview."""
-
-    type: str
-    url: str
-    content: Optional[str] = None
-    mime_type: Optional[str] = None
-    size: int
-    name: str
-
-
-class FileSandboxDeleteResponse(BaseModel):
-    """Response for DELETE /files/delete."""
-
-    message: str
-
-
-class FileSandboxCreateDirResponse(BaseModel):
-    """Response for POST /files/create_directory."""
-
-    message: str
-    directory_info: Any
-
-
-class FileSandboxTreeResponse(BaseModel):
-    """Response for GET /files/tree."""
-
-    path: str
-    tree: List[Any]
-
-
-class FileSandboxStatsResponse(BaseModel):
-    """Response for GET /files/stats."""
-
-    sandbox_root: str
-    total_files: int
-    total_directories: int
-    total_size: int
-    total_size_mb: float
-    max_file_size_mb: int
-    allowed_extensions: List[str]
-
-
-class AdminFileListResponse(BaseModel):
-    """Response for GET /files (admin list directory)."""
-
-    files: List[Any]
-
-
-class AdminFileReadResponse(BaseModel):
-    """Response for GET /files/read (admin read file)."""
-
-    content: str
