+++ autobot-backend/api/terminal_tools.py
@@ -22,7 +22,7 @@ from typing import Any, Dict
 
 from fastapi import APIRouter
 
-from api.schemas_common import PackageManagersResponse
+from api.schemas_terminal import PackageManagersResponse
 from api.terminal_models import ToolInstallRequest
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
